ai-engine/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.26.4
pydantic==2.6.4
Pillow==10.3.0
# Optional — install these for full AI detection
# opencv-python==4.9.0.80
# mediapipe==0.10.11
